Add AppKit font tests and update UIFontBehaviorTests to be BONFontBehaviorTests.
